refactor: GraphView 위치 변경
